End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,13 +17,13 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
   }
 
-  loadMessages(){
+  loadMessages() {
     this.loading = true;
     this.messageService.getMessages(this.pageNumber, this.pageSize, this.container)
       .subscribe({
@@ -35,18 +36,26 @@ export class MessagesComponent implements OnInit {
   }
 
   pageChanged(event: any) {
-    if(this.pageNumber !== event.page) {
+    if (this.pageNumber !== event.page) {
       this.pageNumber = event.page;
       this.loadMessages();
     }
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id)
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone')
       .subscribe({
-        next: () => {
-          this.messages.splice(this.messages.findIndex(m => m.id === id), 1); // 1 is how many messages we want to delete.
+        next: (result) => {
+          if (result) {
+            this.messageService.deleteMessage(id)
+              .subscribe({
+                next: () => {
+                  this.messages.splice(this.messages.findIndex(m => m.id === id), 1); // 1 is how many messages we want to delete.
+                }
+              })
+          }
         }
       })
+
   }
 }
